+++ src/main/java/cn/memoryzy/json/action/test/TestCatNewConfigAction.java
@@ -1,14 +1,12 @@
 package cn.memoryzy.json.action.test;
 
-import cn.memoryzy.json.ui.editor.EditExtension;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.UpdateInBackground;
 import com.intellij.openapi.project.DumbAwareAction;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.DialogBuilder;
-import com.intellij.ui.components.fields.ExtendableTextField;
-import com.intellij.util.ui.JBUI;
-import com.intellij.util.ui.components.BorderLayoutPanel;
+import com.intellij.openapi.wm.StatusBar;
+import com.intellij.openapi.wm.StatusBarWidget;
+import com.intellij.openapi.wm.WindowManager;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -240,18 +238,24 @@ public class TestCatNewConfigAction extends DumbAwareAction implements UpdateInB
         // JavaUtil.hasJavaOrKotlinEnvironment();
 
 
-        ExtendableTextField extendableTextField = new ExtendableTextField(15);
-        extendableTextField.setExtensions(new EditExtension());
+        // ExtendableTextField extendableTextField = new ExtendableTextField(15);
+        // extendableTextField.setExtensions(new EditExtension());
+        //
+        // extendableTextField.setOpaque(false);
+        // extendableTextField.setBorder(JBUI.Borders.empty());
+        //
+        //
+        // BorderLayoutPanel panel = new BorderLayoutPanel().addToCenter(extendableTextField);
+        //
+        // new DialogBuilder()
+        //         .centerPanel(panel)
+        //         .show();
 
-        extendableTextField.setOpaque(false);
-        extendableTextField.setBorder(JBUI.Borders.empty());
 
+        StatusBar statusBar = WindowManager.getInstance().getStatusBar(project);
+        StatusBarWidget widget = statusBar.getWidget("Position");
 
-        BorderLayoutPanel panel = new BorderLayoutPanel().addToCenter(extendableTextField);
-
-        new DialogBuilder()
-                .centerPanel(panel)
-                .show();
+        System.out.println();
 
     }
 }