+++ src/main/java/cn/memoryzy/json/enums/HistoryAffectType.java
@@ -1,0 +1,11 @@
+package cn.memoryzy.json.enums;
+
+/**
+ * @author Memory
+ * @since 2025/8/22
+ */
+public enum HistoryAffectType {
+
+    ADD, UPDATE, REMOVE
+
+}